requirements1.txt
streamlit
faiss-cpu
numpy
langchain
langchain-google-genai
google-generativeai
bs4
python-dotenv
pymupdf